Refactored the Timesheet

+++ apps/frontend/modules/timesheet/actions/actions.class.php
@@ -88,6 +88,8 @@ class timesheetActions extends sfActions
     public function executeUpdate($request)
     {
         $time_values = $request->getParameter('time', array());
+        
+        $user = UserTable::getInstance()->find($this->getUser()->getAttribute('uid'));
 
         $this->time_values = $time_values;
 
@@ -98,38 +100,44 @@ class timesheetActions extends sfActions
             
             $projects = $time_values[$i];
 
-            foreach ($projects as $pid => $project) {
-                $booking_date = date("Y-m-d", $request->getParameter('weekstart') + ($i - 1) * 24 * 60 * 60);
+            $booking_timestamp = $request->getParameter('weekstart') + ($i - 1) * 24 * 60 * 60;
+            $booking_date = date("Y-m-d", $booking_timestamp);
 
-                $query = new Doctrine_Query();
-                $query->delete('TimeLogItem ti')->where('ti.user_id=? AND ti.project_id=? AND ti.itemdate=?',
-                                array($this->getUser()->getAttribute('uid'),
-                                    $pid,
-                                    $booking_date))
-                        ->execute();
+            if ($booking_timestamp < time()) {
+                foreach ($projects as $pid => $project) {
+                    $query = new Doctrine_Query();
+                    $query->delete('TimeLogItem ti')->where('ti.user_id=? AND ti.project_id=? AND ti.itemdate=?',
+                                    array($this->getUser()->getAttribute('uid'),
+                                        $pid,
+                                        $booking_date))
+                            ->execute();
 
-                for ($time_index = 0; $time_index < count($project['time']); $time_index++) {
-                    $time_value = $project['time'][$time_index];
-                    $time_type = $project['type'][$time_index];
-                    $time_comment = $project['comment'][$time_index];
+                    for ($time_index = 0; $time_index < count($project['time']); $time_index++) {
+                        $time_value = $project['time'][$time_index];
+                        $time_type = $project['type'][$time_index];
+                        $time_comment = $project['comment'][$time_index];
 
-                    if (($time_value != "") && ($time_value != 0)) {
-                        $type_query = new Doctrine_Query();
-                        $type = $type_query->from('TimeItemType tit')
-                                        ->where('tit.name=?', array($time_type))
-                                        ->fetchOne();
+                        if (($time_value != "") && ($time_value != 0)) {
+                            $type_query = new Doctrine_Query();
+                            $type = $type_query->from('TimeItemType tit')
+                                            ->where('tit.name=?', array($time_type))
+                                            ->fetchOne();
 
-                        $current_value = new TimeLogItem();
+                            $current_value = new TimeLogItem();
 
-                        $current_value->value = $time_value;
-                        $current_value->itemdate = $booking_date;
-                        $current_value->user_id = $this->getUser()->getAttribute('uid');
-                        $current_value->project_id = $pid;
-                        $current_value->type_id = $type->id;
-                        $current_value->note = $time_comment;
-                        $current_value->save();
+                            $current_value->value = $time_value;
+                            $current_value->itemdate = $booking_date;
+                            $current_value->user_id = $this->getUser()->getAttribute('uid');
+                            $current_value->project_id = $pid;
+                            $current_value->type_id = $type->id;
+                            $current_value->note = $time_comment;
+                            $current_value->save();
+                        }
                     }
                 }
+
+                TimeLogItemTable::getInstance()->updateMissedBookings($booking_timestamp, 
+                                                                        $user);
             }
         }
